backend/requirements.txt
fastapi
uvicorn
pillow
opencv-python
